requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0
apscheduler>=3.10.0
python-telegram-bot>=20.0
rapidfuzz>=3.0.0
lxml>=4.9.0
python-dotenv>=1.0.0
playwright>=1.40.0
